requirements.txt
flask
pandas
requests
gdown
python-dotenv
nltk
scikit-learn
